Introduce routing and refactor App
