+++ package.json
@@ -26,6 +26,7 @@
     "react-konva": "^18.2.12",
     "react-quill": "^2.0.0",
     "react-toastify": "^9.1.3",
+    "react-router-dom": "^6.23.0",
     "uuid": "^11.1.0"
   },
   "devDependencies": {
